+++ docs/src/pages/components/alert/index.page.mdx
@@ -1,6 +1,9 @@
 ---
 title: Alert
 description: An Alert displays a brief, important message in a way that attracts the user's attention without interrupting the user's task. Alerts are typically intended to be read out dynamically by a screen reader.
+ariaPattern: https://www.w3.org/TR/wai-aria-practices/#alert
+themeSource: packages/ui/src/theme/tokens/components/alert.js
+reactSource: packages/react/src/primitives/Alert/Alert.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
