+++ docs/src/pages/components/badge/examples/index.ts
@@ -2,3 +2,4 @@ export { DefaultBadgeExample } from './DefaultBadgeExample';
 export { BadgeVariationExample } from './BadgeVariationExample';
 export { BadgeSizeExample } from './BadgeSizeExample';
 export { BadgeStyleProps } from './BadgeStyleProps';
+export { ThemeExample } from './ThemeExample';
