+++ docs/src/pages/components/badge/index.page.mdx
@@ -1,6 +1,8 @@
 ---
 title: Badge
 description: A Badge is a small visual element to denote a status or message about an item. A small, color-coded visual element that contains letters or numbers, that you can use to label, categorize, or organize items.
+themeSource: packages/ui/src/theme/tokens/components/badge.js
+reactSource: packages/react/src/primitives/Badge/Badge.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
