+++ docs/src/pages/components/button/index.page.mdx
@@ -2,6 +2,8 @@
 title: Button
 description: The Button primitive is used to trigger an action or event, such as submitting a form, opening a dialog, canceling an action, or performing a delete operation.
 ariaPattern: https://www.w3.org/TR/wai-aria-practices/#button
+themeSource: packages/ui/src/theme/tokens/components/button.js
+reactSource: packages/react/src/primitives/Button/Button.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
