+++ docs/src/pages/components/button/react.mdx
@@ -1,10 +1,10 @@
 import { Button, ButtonGroup, Flex, View } from '@aws-amplify/ui-react';
-import { Example, ExampleCode } from '@/components/Example';
-import { ComponentClassTable } from '@/components/ComponentClassTable';
 
-import { ButtonDemo } from './demo';
+import { ComponentClassTable } from '@/components/ComponentClassTable';
 import { Example, ExampleCode } from '@/components/Example';
 import { ComponentThemeTokens } from '@/components/ComponentThemeTokens';
+
+import { ButtonDemo } from './demo';
 import { IconButtonExample, ButtonThemeExample } from './examples';
 
 ## Demo
