+++ docs/src/pages/components/card/index.page.mdx
@@ -1,6 +1,8 @@
 ---
 title: Card
 description: The Card component can be used to group related pieces of content.
+themeSource: packages/ui/src/theme/tokens/components/card.js
+reactSource: packages/react/src/primitives/Card/Card.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
