+++ docs/src/pages/components/checkboxfield/index.page.mdx
@@ -1,5 +1,9 @@
 ---
 title: CheckboxField
+description: CheckboxField is used to mark an individual item as selected, or to select multiple items from a list of individual items.
+ariaPattern: https://www.w3.org/TR/wai-aria-practices/#checkbox
+themeSource: packages/ui/src/theme/tokens/components/checkbox.js
+reactSource: packages/react/src/primitives/CheckboxField/CheckboxField.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
