+++ docs/src/pages/components/checkboxfield/react.mdx
@@ -7,8 +7,6 @@ import { Example } from '@/components/Example';
 import { Fragment } from '@/components/Fragment';
 import { ComponentClassTable } from '@/components/ComponentClassTable';
 
-CheckboxField is used to mark an individual item as selected, or to select multiple items from a list of individual items.
-
 ## Demo
 
 <CheckboxDemo />
