+++ docs/src/pages/components/divider/index.page.mdx
@@ -1,6 +1,8 @@
 ---
 title: Divider
 description: A Divider creates separations in content. Dividers can help organize content and establish visual rhythm.
+themeSource: packages/ui/src/theme/tokens/components/divider.js
+reactSource: packages/react/src/primitives/Divider/Divider.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
