+++ docs/src/pages/components/expander/index.page.mdx
@@ -1,6 +1,9 @@
 ---
 title: Expander
 description: The Expander primitive enables users to expand or collapse a set of sections.
+ariaPattern: https://www.w3.org/TR/wai-aria-practices/#accordion
+themeSource: packages/ui/src/theme/tokens/components/expander.js
+reactSource: packages/react/src/primitives/Expander/Expander.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
