+++ docs/src/pages/components/heading/index.page.mdx
@@ -1,6 +1,8 @@
 ---
 title: Heading
 description: This is a separate primitive from Text for semantic purposes. The Heading primitive maps to an <h*> tag, and Text maps to a <p> tag.
+themeSource: packages/ui/src/theme/tokens/components/heading.js
+reactSource: packages/react/src/primitives/Heading/Heading.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
