+++ docs/src/pages/components/link/index.page.mdx
@@ -1,5 +1,9 @@
 ---
 title: Link
+description: Links are customizable and themable elements used for Navigation. By default Links render an anchor tag but can be configured to be used with routing libraries.
+ariaPattern: https://www.w3.org/TR/wai-aria-practices/#link
+themeSource: packages/ui/src/theme/tokens/components/link.js
+reactSource: packages/react/src/primitives/Link/Link.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
