+++ docs/src/pages/components/link/react.mdx
@@ -3,8 +3,6 @@ import { Example } from '@/components/Example';
 import { ComponentClassTable } from '@/components/ComponentClassTable';
 import { LinkDemo } from './demo';
 
-Links are customizable and themable elements used for Navigation. By default Links render an anchor tag but can be configured to be used with routing libraries.
-
 ## Demo
 
 <LinkDemo />
