+++ docs/src/pages/components/loader/index.page.mdx
@@ -1,5 +1,8 @@
 ---
 title: Loader
+description: Loaders provide a visual cue that an action is either processing or awaiting a result. They are used to help the customer understand the system is working to fulfill a request.
+themeSource: packages/ui/src/theme/tokens/components/loader.js
+reactSource: packages/react/src/primitives/Loader/Loader.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
