+++ docs/src/pages/components/loader/react.mdx
@@ -4,8 +4,6 @@ import { LoaderDemo } from './demo';
 import { Example } from '@/components/Example';
 import { ComponentClassTable } from '@/components/ComponentClassTable';
 
-Loaders provide a visual cue that an action is either processing or awaiting a result. They are used to help the customer understand the system is working to fulfill a request.
-
 ## Demo
 
 <LoaderDemo />
