+++ docs/src/pages/components/menu/index.page.mdx
@@ -1,8 +1,9 @@
 ---
 title: Menu
+description: Menu provides an accessible, interactive menu for selecting actions within an application. Dropdown menu is collision-aware and will automatically change location based on available space.
 ariaPattern: https://www.w3.org/TR/wai-aria-practices-1.2/#menubutton
-themeSource: packages/ui/src/theme/tokens/components/button.js
-reactSource: packages/react/src/Button/Button.tsx
+themeSource: packages/ui/src/theme/tokens/components/menu.js
+reactSource: packages/react/src/primitives/Menu/Menu.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
