+++ docs/src/pages/components/menu/react.mdx
@@ -12,8 +12,6 @@ import {
 import { Example } from '@/components/Example';
 import { ComponentClassTable } from '@/components/ComponentClassTable';
 
-Menu provides an accessible, interactive menu for selecting actions within an application. Dropdown menu is collision-aware and will automatically change location based on available space.
-
 ## Demo
 
 <MenuDemo />
