+++ docs/src/pages/components/pagination/index.page.mdx
@@ -1,5 +1,8 @@
 ---
 title: Pagination
+description: Pagination provides navigation to allow customers to move between large sets of content that are distributed across multiple pages.
+themeSource: packages/ui/src/theme/tokens/components/pagination.js
+reactSource: packages/react/src/primitives/Pagination/Pagination.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
