+++ docs/src/pages/components/pagination/react.mdx
@@ -3,8 +3,6 @@ import { PaginationDemo } from './demo.tsx';
 import { Example } from '@/components/Example';
 import { ComponentClassTable } from '@/components/ComponentClassTable';
 
-Pagination provides navigation to allow customers to move between large sets of content that are distributed across multiple pages.
-
 ## Demo
 
 <PaginationDemo />
