+++ docs/src/pages/components/passwordfield/index.page.mdx
@@ -1,5 +1,7 @@
 ---
 title: PasswordField
+description: The PasswordField primitive allows users to input passwords. It also features full password manager support and an optional show/hide password button for user convenience.
+reactSource: packages/react/src/primitives/PasswordField/PasswordField.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
