+++ docs/src/pages/components/passwordfield/react.mdx
@@ -19,8 +19,6 @@ import { Fragment } from '@/components/Fragment';
 import { ComponentClassTable } from '@/components/ComponentClassTable';
 import { RefExample } from './examples/refs';
 
-The PasswordField primitive allows users to input passwords. It also features full password manager support and an optional show/hide password button for user convenience.
-
 ## Demo
 
 <PasswordFieldDemo />
