+++ docs/src/pages/components/phonenumberfield/index.page.mdx
@@ -1,5 +1,7 @@
 ---
 title: PhoneNumberField
+description: The PhoneNumberField form primitive can be used allow users to input a phone number.
+reactSource: packages/react/src/primitives/PhoneNumberField/PhoneNumberField.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
