+++ docs/src/pages/components/phonenumberfield/react.mdx
@@ -18,8 +18,6 @@ import { Fragment } from '@/components/Fragment';
 import { ComponentClassTable } from '@/components/ComponentClassTable';
 import { RefExample } from './examples/refs';
 
-The `PhoneNumberField` form primitive can be used allow users to input a phone number.
-
 ## Demo
 
 <PhoneNumberFieldDemo />
