+++ docs/src/pages/components/placeholder/index.page.mdx
@@ -1,5 +1,8 @@
 ---
 title: Placeholder
+description: The Placeholder component is used to fill out the interface while content is loaded asynchronously.
+themeSource: packages/ui/src/theme/tokens/components/placeholder.js
+reactSource: packages/react/src/primitives/Placeholder/Placeholder.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
