+++ docs/src/pages/components/placeholder/react.mdx
@@ -3,8 +3,6 @@ import { PlaceholderDemo, TwoWays } from './demo';
 import { Example } from '@/components/Example';
 import { ComponentClassTable } from '@/components/ComponentClassTable';
 
-The Placeholder component is used to fill out the interface while content is loaded asynchronously.
-
 ## Demo
 
 <PlaceholderDemo />
