+++ docs/src/pages/components/togglebutton/index.page.mdx
@@ -1,6 +1,9 @@
 ---
 title: ToggleButton
 description: A toggle button represents an on/off state for some configuration, for example switching on/off bold text in a text editor.
+ariaPattern: https://www.w3.org/TR/wai-aria-practices-1.2/#button
+themeSource: packages/ui/src/theme/tokens/components/toggleButton.js
+reactSource: packages/react/src/primitives/ToggleButton/ToggleButton.tsx
 ---
 
 import { Fragment } from '@/components/Fragment';
