+++ packages/react/src/index.tsx
@@ -10,5 +10,10 @@ export * as primitives from './primitives';
 /**
  * Re-export public APIs from `@aws-amplify/ui`
  */
-export { defaultTheme, createTheme, translations } from '@aws-amplify/ui';
+export {
+  baseTheme,
+  defaultTheme,
+  createTheme,
+  translations,
+} from '@aws-amplify/ui';
 export type { Theme } from '@aws-amplify/ui';
